Add missing Apache License headers
Per license recommendation:

To apply the Apache License to your work, attach the license
boilerplate notice, The text should be enclosed in the appropriate
comment syntax for the file format. We also recommend that a file or
class name and description of purpose be included on the same
"printed page" as the copyright notice for easier identification
within third-party archives.