Remove 4 minute timeout for log follow process (#107)
* Remove 4 minute timeout for log follow process (#106)

The process that follows the training logs of an ongoing training job
should not timeout after 4 minutes. Instead the log follow process
should complete after the training job itself is finished.

This behavior is necessary to enable chaining up commands to create
machine learning pipelines, where subsequent commands require the output
data of the training job whose logs are being "followed" like in our
ART notebook.

This commit reinstates the log follow behavior prior merge of PR #79

* Updates suggested by sboagibm

Intention was to not rely on a long term stream being held open, 
but to be able to re-open a new stream starting from where the 
old left off, if the connection terminates.